fix: force relative URL resolution;
When using the `data-main` attribute, Chrome(72) WILL NOT assume that the initial request is from the document’s origin. Instead, it will construct a URL relative to the asset’s origin. For example:

<script src=“//unpkg.com/dimport” data-main=“/bundle.js” />

In Chrome and Safari, and without this fix, dimport’s first request is to “https://unpkg.com/bundle.js” but in Firefox, the first request is correctly made to “localhost/bundle.js”